+++ db/migrate/20160808081300_add_external_id_to_txns.rb
@@ -1,0 +1,6 @@
+class AddExternalIdToTxns < ActiveRecord::Migration
+  def change
+    add_column :bloom_remit_txns, :external_id, :string, null: false
+    add_index :bloom_remit_txns, :external_id
+  end
+end