+++ spec/concepts/bloom_remit/txns/operations/create_spec.rb
@@ -8,6 +8,7 @@ module BloomRemit
         let(:user) { create(:bloom_remit_dummy_sub_user) }
         let(:biller) { create(:bloom_remit_dummy_biller, slug: "PLDT") }
         let(:payment) { create(:bloom_remit_dummy_sub_payment) }
+        let(:external_id) { SecureRandom.hex(12) }
 
         it "creates a txn with a secret and enqueues the PayoutJob" do
           op = described_class.(txn: {
@@ -19,6 +20,7 @@ module BloomRemit
             account_id: "Hooli X",
             owner_id: payment.id,
             owner_type: payment.class.base_class.name,
+            external_id: external_id,
           })
           txn = op.model
           expect(txn.target_slug).to eq biller.slug
@@ -27,6 +29,7 @@ module BloomRemit
           expect(txn.account_name).to eq "028109090"
           expect(txn.account_id).to eq "Hooli X"
           expect(txn.owner).to eq payment
+          expect(txn.external_id).to eq external_id
           expect(PayoutJob).to have_enqueued_job(txn.id)
         end